requirements.txt
biopython==1.76
click==7.1.1
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.15.1
MarkupSafe==1.1.1
mysql-connector==2.2.9
nltk==3.5
numpy==1.18.4
pkg-resources==0.0.0
regex==2020.5.14
tqdm==4.46.0
Werkzeug==1.0.1
